+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Terasology Launcher - ChangeLog
 ===============================
 
 ## x.y.z (unreleased)
+* Translation into Russian
 
 ## 1.2.0 (2014-01-17)
 * Choose, edit, show and use game data directory